Always update heart/cassette count
Autosplitter needs to always be aware of a change in collectible count when it occurs and so that it doesn't misfire on future splits. Hard to explain succinctly but e.g. collecting a B-side heart while on a chapter complete autosplit means that the heart is never accounted for, so if the next two autosplits are a checkpoint and then a heart collect, as soon as the checkpoint fires the heart collect fires. This occurs in All Hearts runs between 3B and 4A for example.